v2.3
Make Pear run optionally
Add SPAdes 3.10.1
Save reports in json (optional)
Get maximum reads length from fastQintegrity for downstream modules
indenpendency
Save excluded contigs (optional)
Save assembly mapping coverage before and after filtration
Assembly mapping Qa/QC based on coverage after filtration
Gene presence/abscence in trueCoverage relies also in gene identity
In trueCoverage remove bases with Phred score lower than 7
Add Yersinia enterocolitica trueCoverage files set and Streptococcus
agalactiae trueCoverage files set (the latter based on MLST genes)
Change subprocess timeout kill
Change programs version check
Update README

+++ modules/fastqc.py
@@ -2,6 +2,7 @@ import sys
 import os
 import utils
 from functools import partial
+import time
 
 
 # Prepare adapters fasta file to FastQC (tabular file)
@@ -53,6 +54,7 @@ def parseFastQC(fastqc_folder, fastq_files):
 	goodReads = []
 	badReads = []
 	failing = {}
+	warning = {}
 	for reads in fastq_files:
 		reads_file = reads.rsplit('.', 2)[0]
 		with open(os.path.join(fastqc_folder, str(reads_file + '_fastqc'), 'summary.txt'), 'rtU') as fastqc_summary:
@@ -66,9 +68,6 @@ def parseFastQC(fastqc_folder, fastq_files):
 			if fastqc.get('Per base sequence quality') != 'PASS':
 				bad_fastq = True
 				failing[reads].append('Bad per base sequence quality')
-			if fastqc.get('Per base sequence content') == 'FAIL':
-				bad_fastq = True
-				failing[reads].append('Bad per base sequence content')
 			if fastqc.get('Per sequence GC content') == 'FAIL':
 				bad_fastq = True
 				failing[reads].append('Bad per sequence GC content')
@@ -85,6 +84,11 @@ def parseFastQC(fastqc_folder, fastq_files):
 				bad_fastq = True
 				failing[reads].append('Found adapters sequences')
 
+			if fastqc.get('Per base sequence content') == 'FAIL':
+				if reads not in warning:
+					warning[reads] = []
+				warning[reads].append('Bad per base sequence content')
+
 			if not bad_fastq:
 				goodReads.append(reads)
 			else:
@@ -94,7 +98,7 @@ def parseFastQC(fastqc_folder, fastq_files):
 		if len(failing[fastq]) == 0:
 			failing[fastq] = False
 
-	return goodReads, badReads, failing
+	return goodReads, badReads, failing, warning
 
 
 # Get reads length data, nucleotide bias status & number of reads
@@ -233,16 +237,18 @@ fastqc_timer = partial(utils.timer, name='FastQC analysis')
 
 # Run FastQC analysis
 @fastqc_timer
-def runFastQCanalysis(outdir, threads, adaptersFasta, fastq_files):
+def runFastQCanalysis(outdir, threads, adaptersFasta, fastq_files, keepFiles, fastQC_run_name):
 	pass_qc = False
 	failing = {}
 	failing['sample'] = False
+
+	warnings = {}
 
 	maximumReadsLength = None
 	nts2clip_based_ntsContent = None
 
 	# Create FastQC output directory
-	fastqc_folder = os.path.join(outdir, 'fastqc', '')
+	fastqc_folder = os.path.join(outdir, str('fastqc_' + fastQC_run_name), '')
 	utils.removeDirectory(fastqc_folder)
 	os.mkdir(fastqc_folder)
 
@@ -256,7 +262,7 @@ def runFastQCanalysis(outdir, threads, adaptersFasta, fastq_files):
 			return run_successfully, pass_qc, failing, maximumReadsLength, nts2clip_based_ntsContent
 
 		# Check which reads pass FastQC
-		goodReads, badReads, failing = parseFastQC(fastqc_folder, fastq_files)
+		goodReads, badReads, failing, warnings = parseFastQC(fastqc_folder, fastq_files)
 		# Get reads information
 		maximumReadsLength, moreFrequentReadsLength, numberReads, ntsContent_biasStatus = getReadsInformation(fastqc_folder, fastq_files)
 		# Get number nucleotides to clip based on nucleotide content bias
@@ -276,6 +282,7 @@ def runFastQCanalysis(outdir, threads, adaptersFasta, fastq_files):
 		failing['sample'] = 'Did not run'
 		print failing['sample']
 
-	utils.removeDirectory(fastqc_folder)
+	if not keepFiles:
+		utils.removeDirectory(fastqc_folder)
 
-	return run_successfully, pass_qc, failing, maximumReadsLength, nts2clip_based_ntsContent
+	return run_successfully, pass_qc, failing, warnings, maximumReadsLength, nts2clip_based_ntsContent
